+++ src/cjs/draw.d.ts
@@ -60,6 +60,7 @@
  * @modified 2026-01-04 Adding `lineJoin` attribute to the `StrokeOptions`.
  * @modified 2026-03-18 Adding `isOpen` parameter to `cubicBezierPath` draw method.
  * @modified 2026-04-04 Added the method `bounds`.
+ * @modified 2026-04-04 Handling the `lineCap` attribute in the `StrokeOptions`.
  * @version  1.15.0
  **/
 import { Polygon } from "./Polygon";