+++ src/cjs/drawutilssvg.js
@@ -53,6 +53,7 @@
  * @modified 2026-01-04 Fixing missing `strokeOptions` param in the `drawutilssvg.polygon` method.
  * @modified 2026-03-18 Adding `isOpen` parameter to `cubicBezierPath` draw method.
  * @modified 2026-04-04 Added the method `bounds`.
+ * @modified 2026-04-04 Handling the `stroke-linecap` option now from the `StrokeOptions` interface.
  * @version  1.7.0
  **/
 Object.defineProperty(exports, "__esModule", { value: true });
@@ -478,6 +479,10 @@ var drawutilssvg = /** @class */ (function () {
         // Set line join option?
         if (strokeOptions && strokeOptions.lineJoin && drawutilssvg.nodeSupportsLineJoin(node.tagName)) {
             node.setAttribute("stroke-linejoin", strokeOptions.lineJoin);
+        }
+        // Set lineCap option?
+        if (strokeOptions && strokeOptions.lineCap && drawutilssvg.nodeSupportsLineCap(node.tagName)) {
+            node.setAttribute("stroke-linecap", strokeOptions.lineCap);
         }
     };
     drawutilssvg.prototype._x = function (x) {
@@ -1598,6 +1603,9 @@ var drawutilssvg = /** @class */ (function () {
         return ["line", "path", "circle", "ellipse", "rectangle", "rect"].includes(nodeName);
     };
     drawutilssvg.nodeSupportsLineJoin = function (nodeName) {
+        return ["line", "path", "circle", "ellipse", "rectangle", "rect"].includes(nodeName);
+    };
+    drawutilssvg.nodeSupportsLineCap = function (nodeName) {
         return ["line", "path", "circle", "ellipse", "rectangle", "rect"].includes(nodeName);
     };
     /**